refactor: moved signing stuff around
