+++ Cargo.toml
@@ -148,7 +148,7 @@ iam-rs = { version = "=0.0.23" } # AWS IAM policy parser/generator - HIGH RISK: 
 itertools = "=0.14.0"  # Iterator utilities - LOW RISK: bluss/rust-itertools, 100M+ downloads
 k256 = "=0.13.4"  # secp256k1 elliptic curve - LOW RISK: RustCrypto org, 33M+ downloads
 minijinja = { version = "=2.20.0", features = ["loader"] }  # Template engine - HIGH RISK: Individual maintainer (mitsuhiko), despite exceptional track record
-ml-dsa = { version = "=0.1.1", default-features = false, features = ["alloc", "zeroize"] }  # ML-DSA (FIPS 204) post-quantum signatures for hybrid PQ signing (issue #3078) - MEDIUM RISK: RustCrypto org, new standard implementation, but no security audit
+ml-dsa = { version = "=0.1.1", default-features = false, features = ["alloc", "zeroize"] }  # ML-DSA (FIPS 204) post-quantum signatures for hybrid PQ signing - MEDIUM RISK: RustCrypto org, new standard implementation, but no security audit
 ml-kem = { version = "=0.2.3", features = ["zeroize"] }  # ML-KEM (Kyber) post-quantum KEM - MEDIUM RISK: New standard implementation, needs security audit
 mockall = "=0.13.1"  # Mocking for tests - HIGH RISK: Individual maintainer (asomers), test-only dependency
 ndarray = { version = "=0.16.1", features = ["serde"] }  # N-dimensional arrays - LOW RISK: rust-ndarray team
