+++ core/service/Cargo.toml
@@ -62,7 +62,7 @@ cfg-if.workspace = true
 ciborium.workspace = true
 clap = { workspace = true, features = ["derive"] }
 console_error_panic_hook.workspace = true
-ed25519-dalek = { workspace = true, optional = true }
+ed25519-dalek.workspace = true
 enum_dispatch.workspace = true
 futures-util.workspace = true
 hashing.workspace = true
@@ -71,7 +71,7 @@ iam-rs = { workspace = true, optional = true }
 itertools.workspace = true
 k256.workspace = true
 kms-grpc = { workspace = true, default-features = false }
-ml-dsa = { workspace = true, optional = true }
+ml-dsa.workspace = true
 ml-kem.workspace = true
 nom.workspace = true
 nsm-nitro-enclave-utils = { version = "0.1.2", features = ["pki"], optional = true }
@@ -157,8 +157,6 @@ non-wasm = [
     "threshold-execution/non-wasm",
     "dep:threshold-networking",
     "dep:attestation-doc-validation",
-    "dep:ed25519-dalek",
-    "dep:ml-dsa",
     "dep:aws-config",
     "dep:aws-nitro-enclaves-nsm-api",
     "dep:aws-sdk-kms",
