+++ core/service/src/cryptography/mod.rs
@@ -5,10 +5,7 @@ pub mod error;
 pub mod hybrid_ml_kem;
 pub mod internal_crypto_types;
 pub mod signatures;
-#[cfg(feature = "non-wasm")]
 pub mod signing;
-#[cfg(feature = "non-wasm")]
-pub mod signing_schemes;
 // Allow our deprecated modules for now as we need to be backwards compatible
 #[allow(deprecated)]
 pub mod encryption;
